receiver in server

+++ 05ass/message.h
@@ -21,9 +21,9 @@ struct Message {
   Message(const std::string &tag, const std::string &data)
     : tag(tag), data(data) { }
 
-  vector<string> format_data() {
+  vector<string> format_data() { // split message string formatted as room:sender:message
     vector<string> formatted_data;
-    int colon = data.find(':');
+    size_t colon = data.find(':');
     while (colon != string::npos) {
       formatted_data.push_back(data.substr(0, colon));
       data.erase(0, colon+1);
